# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "AdminSeriesCreatedResponse": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug"
        ],
        "title": "AdminSeriesCreatedResponse",
        "type": "object"
      },
      "AdminSeriesDetailResponse": {
        "properties": {
          "accentColor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accentcolor"
          },
          "autoStartNextRound": {
            "title": "Autostartnextround",
            "type": "boolean"
          },
          "coverImageUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverimageurl"
          },
          "defaultPolicies": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Defaultpolicies",
            "type": "array"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "groups": {
            "items": {
              "$ref": "#/components/schemas/AdminGroupResponse"
            },
            "title": "Groups",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "isArchived": {
            "title": "Isarchived",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "roundPlan": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roundplan"
          },
          "rounds": {
            "items": {
              "$ref": "#/components/schemas/AdminRoundResponse"
            },
            "title": "Rounds",
            "type": "array"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "description",
          "timezone",
          "defaultPolicies",
          "roundPlan",
          "autoStartNextRound",
          "isArchived",
          "coverImageUrl",
          "accentColor",
          "groups",
          "rounds"
        ],
        "title": "AdminSeriesDetailResponse",
        "type": "object"
      },
      "AdminSeriesResponse": {
        "properties": {
          "accentColor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accentcolor"
          },
          "autoStartNextRound": {
            "title": "Autostartnextround",
            "type": "boolean"
          },
          "coverImageUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverimageurl"
          },
          "defaultPolicies": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Defaultpolicies",
            "type": "array"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "isArchived": {
            "title": "Isarchived",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "roundPlan": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roundplan"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "description",
          "timezone",
          "defaultPolicies",
          "roundPlan",
          "autoStartNextRound",
          "isArchived",
          "coverImageUrl",
          "accentColor"
        ],
        "title": "AdminSeriesResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SeriesCreate": {
        "properties": {
          "accent_color": {
            "anyOf": [
              {
                "pattern": "^#[0-9a-fA-F]{6}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accent Color"
          },
          "auto_start_next_round": {
            "default": true,
            "title": "Auto Start Next Round",
            "type": "boolean"
          },
          "cover_image_url": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Image Url"
          },
          "default_policies": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Default Policies",
            "type": "array"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "round_plan": {
            "anyOf": [
              {
                "discriminator": {
                  "mapping": {
                    "calendar": "#/components/schemas/CalendarRoundPlan",
                    "rolling": "#/components/schemas/RollingRoundPlan"
                  },
                  "propertyName": "kind"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/RollingRoundPlan"
                  },
                  {
                    "$ref": "#/components/schemas/CalendarRoundPlan"
                  }
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Round Plan"
          },
          "slug": {
            "maxLength": 100,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
            "title": "Slug",
            "type": "string"
          },
          "timezone": {
            "default": "UTC",
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "name",
          "slug"
        ],
        "title": "SeriesCreate",
        "type": "object"
      },
...
      "SeriesUpdate": {
        "description": "Mutable series defaults; existing materialized rounds are not rewritten.",
        "properties": {
          "accent_color": {
            "anyOf": [
              {
                "pattern": "^#[0-9a-fA-F]{6}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accent Color"
          },
          "auto_start_next_round": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Start Next Round"
          },
          "cover_image_url": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Image Url"
          },
          "default_policies": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Policies"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 10000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_archived": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Archived"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "round_plan": {
            "anyOf": [
              {
                "discriminator": {
                  "mapping": {
                    "calendar": "#/components/schemas/CalendarRoundPlan",
                    "rolling": "#/components/schemas/RollingRoundPlan"
                  },
                  "propertyName": "kind"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/RollingRoundPlan"
                  },
                  {
                    "$ref": "#/components/schemas/CalendarRoundPlan"
                  }
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Round Plan"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          }
        },
        "title": "SeriesUpdate",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "BoroCrew Music API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/series": {
      "get": {
        "operationId": "list_series_for_administration_api_v1_admin_series_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AdminSeriesResponse"
                  },
                  "title": "Response List Series For Administration Api V1 Admin Series Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Series For Administration",
        "tags": [
          "administration"
        ]
      },
      "post": {
        "operationId": "create_series_api_v1_admin_series_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeriesCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSeriesCreatedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Series",
        "tags": [
          "administration"
        ]
      }
    },
    "/api/v1/admin/series/{series_id}": {
      "get": {
        "operationId": "get_series_for_administration_api_v1_admin_series__series_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "series_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Series Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSeriesDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Series For Administration",
        "tags": [
          "administration"
        ]
      },
      "patch": {
        "description": "Change future-series defaults without changing existing round snapshots.",
        "operationId": "update_series_api_v1_admin_series__series_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "series_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Series Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeriesUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSeriesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Series",
        "tags": [
          "administration"
        ]
      }
    },
...
    "/api/v1/series/invites/{token}/accept": {
      "post": {
        "operationId": "accept_series_invite_api_v1_series_invites__token__accept_post",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Accept Series Invite Api V1 Series Invites  Token  Accept Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Accept Series Invite",
        "tags": [
          "series"
        ]
      }
    },
    "/api/v1/series/{series_id}": {
      "get": {
        "description": "Return rounds visible to the caller within one series.\n\nSeries may run overlapping contributor groups. A normal contributor sees\nonly their active materialized memberships; platform and series admins can\ninspect the complete series history.",
        "operationId": "get_series_history_api_v1_series__series_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "series_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Series Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesHistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Series History",
        "tags": [
          "series"
        ]
      }
    }
  }
}